added redux slice for user details
